+++ client-v2/src/main/java/com/clickhouse/client/api/data_formats/internal/SerializerUtils.java
@@ -558,10 +558,10 @@ public class SerializerUtils {
                 BinaryStreamUtils.writeBFloat16(stream, (float) value);
                 break;
             case Float32:
-                BinaryStreamUtils.writeFloat32(stream, (float) value);
+                BinaryStreamUtils.writeFloat32(stream, NumberConverter.toFloat(value));
                 break;
             case Float64:
-                BinaryStreamUtils.writeFloat64(stream, (double) value);
+                BinaryStreamUtils.writeFloat64(stream, NumberConverter.toDouble(value));
                 break;
             case Decimal:
             case Decimal32: